Puppetize Horizon element
Converts the Horizon element to use puppet instead of git.

django_debug is set because I haven't found a way to set
ALLOWED_HOSTS in the local_settings, which is required to use a
non-debug configuration in current Django.

Also, the horizon secret_key is currently unset.  I need to find
a better way to handle that.